+++ docs/changelog.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+- Create command-line options for each config option (#373)
+- Overhaul treatment of function definitions (#372)
+  - Support positional-only arguments
+  - Infer more precise types for lambda functions
+  - Infer more precise types for nested functions
+  - Refactor related code
+- Add check for incompatible overrides in child classes
+  (#371)
 - Add `pyanalyze.extensions.NoReturnGuard` (#370)
 - Infer call signatures for `Type[X]` (#369)
 - Support configuration in a `pyproject.toml` file (#368)